After giving this another go I tried having the default methods check for if myQ is an array.
If it's not then it'll do what the method is supposed to do (at least I think). using .isArray(); it will check if myQ is an array.

If it is it will run through the method for the array, .isArray(); returns false then it'll run the method for the list. At least that's what I think

The only issue is that I had to create myQ as an array and a list in the interface, however, they're public but not static (or else we wouldn't be able to edit them).

Hopefully you can see where I'm going with this, but if I'm honest with you my interface knowledge is a little iffy because I am still reading through them and trying to ensure I have a good grasp on them.

The issue is the interface methods do not work with the main method because of the fact that the interface has totally different myQ's than what the main method is reading.